# [Enhancement] Improve deletion vector performance using roaring bitmap instead of set