--- a/be/src/connector/deletion_vector/deletion_vector.h
+++ b/be/src/connector/deletion_vector/deletion_vector.h
@@ -20,13 +20,17 @@
 
 namespace starrocks {
 
+struct DeletionVectorBuildStats {
+    int64_t bitmap_deserialized_ns = 0;
+};
+
 class DeletionVector {
 public:
     DeletionVector(const HdfsScannerParams& scanner_params)
             : _deletion_vector_descriptor(scanner_params.deletion_vector_descriptor), _params(scanner_params) {}
 
-    Status fill_row_indexes(std::set<int64_t>* need_skip_rowids);
-    Status deserialized_inline_dv(std::string& encoded_bitmap_data, std::set<int64_t>* need_skip_rowids) const;
+    Status fill_row_indexes(const SkipRowsContextPtr& skip_rows_ctx);
+    Status deserialized_inline_dv(std::string& encoded_bitmap_data, const SkipRowsContextPtr& skip_rows_ctx);
     StatusOr<std::string> get_absolute_path(const std::string& table_location) const;
 
     const bool is_inline() {
@@ -46,8 +50,8 @@ private:
             std::shared_ptr<io::SharedBufferedInputStream>& shared_buffered_input_stream,
             std::shared_ptr<io::CacheInputStream>& cache_input_stream) const;
 
-    Status deserialized_deletion_vector(uint32_t magic_number, std::unique_ptr<char[]> serialized_dv,
-                                        int64_t serialized_bitmap_length, std::set<int64_t>* need_skip_rowids) const;
+    Status deserialized_deletion_vector(uint32_t magic_number, std::vector<char>& serialized_dv,
+                                        int64_t serialized_bitmap_length, const SkipRowsContextPtr& skip_rows_ctx);
 
     std::string assemble_deletion_vector_path(const std::string& table_location, std::string&& uuid,
                                               std::string& prefix) const;
@@ -57,7 +61,10 @@ private:
                                    const std::shared_ptr<io::CacheInputStream>& cache_input_stream,
                                    const std::shared_ptr<io::SharedBufferedInputStream>& shared_buffered_input_stream);
 
+    void update_dv_build_counter(RuntimeProfile* parent_profile, const DeletionVectorBuildStats& build_stats);
+
     const std::shared_ptr<TDeletionVectorDescriptor> _deletion_vector_descriptor;
     const HdfsScannerParams& _params;
+    DeletionVectorBuildStats _build_stats;
 };
 } // namespace starrocks