--- a/be/src/exec/hdfs_scanner_orc.h
+++ b/be/src/exec/hdfs_scanner_orc.h
@@ -27,7 +27,7 @@ class OrcRowReaderFilter;
 
 class HdfsOrcScanner final : public HdfsScanner {
 public:
-    HdfsOrcScanner() = default;
+    HdfsOrcScanner() : _skip_rows_ctx(std::make_shared<SkipRowsContext>()){};
     ~HdfsOrcScanner() override = default;
 
     Status do_open(RuntimeState* runtime_state) override;
@@ -67,7 +67,7 @@ private:
     std::shared_ptr<OrcRowReaderFilter> _orc_row_reader_filter;
     Filter _dict_filter;
     Filter _chunk_filter;
-    std::set<int64_t> _need_skip_rowids;
+    SkipRowsContextPtr _skip_rows_ctx;
     std::unique_ptr<ORCHdfsFileStream> _input_stream;
 };