--- a/be/src/exec/hdfs_scanner_parquet.cpp
+++ b/be/src/exec/hdfs_scanner_parquet.cpp
@@ -29,8 +29,8 @@ static const std::string kParquetProfileSectionPrefix = "Parquet";
 Status HdfsParquetScanner::do_init(RuntimeState* runtime_state, const HdfsScannerParams& scanner_params) {
     if (!scanner_params.deletes.empty()) {
         SCOPED_RAW_TIMER(&_app_stats.iceberg_delete_file_build_ns);
-        auto iceberg_delete_builder = std::make_unique<IcebergDeleteBuilder>(&_skip_rows_ctx->need_skip_rowids,
-                                                                             runtime_state, scanner_params);
+        auto iceberg_delete_builder =
+                std::make_unique<IcebergDeleteBuilder>(_skip_rows_ctx, runtime_state, scanner_params);
         for (const auto& delete_file : scanner_params.deletes) {
             if (delete_file->file_content == TIcebergFileContent::POSITION_DELETES) {
                 RETURN_IF_ERROR(iceberg_delete_builder->build_parquet(*delete_file));
@@ -44,7 +44,7 @@ Status HdfsParquetScanner::do_init(RuntimeState* runtime_state, const HdfsScanne
         _app_stats.iceberg_delete_files_per_scan += scanner_params.deletes.size();
     } else if (scanner_params.paimon_deletion_file != nullptr) {
         std::unique_ptr<PaimonDeleteFileBuilder> paimon_delete_file_builder(
-                new PaimonDeleteFileBuilder(scanner_params.fs, &_skip_rows_ctx->need_skip_rowids));
+                new PaimonDeleteFileBuilder(scanner_params.fs, _skip_rows_ctx));
         RETURN_IF_ERROR(paimon_delete_file_builder->build(scanner_params.paimon_deletion_file.get()));
     } else if (scanner_params.deletion_vector_descriptor != nullptr) {
         if (scanner_params.split_context != nullptr) {
@@ -54,7 +54,7 @@ Status HdfsParquetScanner::do_init(RuntimeState* runtime_state, const HdfsScanne
         }
         SCOPED_RAW_TIMER(&_app_stats.deletion_vector_build_ns);
         std::unique_ptr<DeletionVector> dv = std::make_unique<DeletionVector>(scanner_params);
-        RETURN_IF_ERROR(dv->fill_row_indexes(&_skip_rows_ctx->need_skip_rowids));
+        RETURN_IF_ERROR(dv->fill_row_indexes(_skip_rows_ctx));
         _app_stats.deletion_vector_build_count += 1;
     }
     return Status::OK();