… be/src/connector/CMakeLists.txt
… be/src/exec/paimon/paimon_delete_file_builder.cpp
… be/src/formats/orc/orc_chunk_reader.cpp
… be/src/formats/orc/orc_chunk_reader.h
… be/src/formats/parquet/file_reader.cpp
… be/src/formats/parquet/group_reader.cpp
… be/src/formats/parquet/group_reader.h
… be/test/connector/deletion_vector/deletion_vector_test.cpp
… be/test/exec/paimon/paimon_delete_file_builder_test.cpp
… be/test/formats/orc/orc_chunk_reader_test.cpp
… be/test/formats/parquet/file_reader_test.cpp
… be/test/formats/parquet/group_reader_test.cpp
… test/sql/test_iceberg/R/test_iceberg_v2
… test/sql/test_iceberg/T/test_iceberg_v2