## Why I'm doing:
Query delta lake table with deletion vector takes lots of time because of building skip rowid set
## What I'm doing:
1. do not build skip row id set, using roaing bitmap to replace set
2. add some meric of deserialized deletion vector

## performance

<meta charset="utf-8"><div data-page-id="KlrddwSRGoDeQHxnPTFc3ScWnJg" data-lark-html-role="root" data-docx-has-block-data="true"><div>
 deletion percent | 0 percent | 5 percent | 10 percent | 30 percent | 50 percent
-- | -- | -- | -- | -- | --
before | 4.9s | 6.5s | 8.6s | 16.4s | 23.1s
with PR 54847 | 4.3s | 5.2s | 5.3s | 5.8s | 5.6s


## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [x] 3.4
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0