@dirtysalt 2025-01-09
why roraing bitmap has better performance than std::set?  the lookup performance is better?